feat: add navbar navigation
