+++ packages/studio-web/src/app/app.component.sass
@@ -7,3 +7,9 @@
   align-items: center
   justify-content: space-between
   align-content:space-around
+
+.nav__button
+  align-items: right
+
+.nav-spacer
+  flex: 1 1 auto
