+++ packages/studio-web/src/app/app.component.ts
@@ -1,7 +1,8 @@
-import { Subject } from "rxjs";
-
+import { Subject, takeUntil } from "rxjs";
+import { MatDialogRef, MatDialog } from "@angular/material/dialog";
 import { Component, OnDestroy, OnInit } from "@angular/core";
 import { environment } from "../environments/environment";
+import { Router } from "@angular/router";
 @Component({
   selector: "app-root",
   templateUrl: "./app.component.html",
@@ -10,8 +11,25 @@ import { environment } from "../environments/environment";
 export class AppComponent implements OnDestroy, OnInit {
   unsubscribe$ = new Subject<void>();
   version = environment.packageJson.singleFileBundleVersion;
-  constructor() {}
-  ngOnInit(): void {}
+  currentURL = "/";
+  constructor(
+    private dialog: MatDialog,
+    public router: Router,
+  ) {}
+  ngOnInit(): void {
+    this.router.events.pipe(takeUntil(this.unsubscribe$)).subscribe((event) => {
+      if (event.type === 1) {
+        this.currentURL = event.url;
+      }
+    });
+  }
+
+  openPrivacyDialog(): void {
+    this.dialog.open(PrivacyDialog, {
+      width: "50vw",
+      maxHeight: "90vh",
+    });
+  }
 
   ngOnDestroy(): void {
     this.unsubscribe$.next();
@@ -19,4 +37,27 @@ export class AppComponent implements OnDestroy, OnInit {
   }
 
   ngAfterViewInit() {}
+}
+
+@Component({
+  selector: "privacy-dialog",
+  templateUrl: "privacy-dialog.html",
+})
+export class PrivacyDialog {
+  analyticsExcluded =
+    window.localStorage.getItem("plausible_ignore") === "true";
+  constructor(public dialogRef: MatDialogRef<PrivacyDialog>) {}
+  ngOnInit() {
+    this.dialogRef.updateSize("100%");
+  }
+
+  toggleAnalytics() {
+    if (this.analyticsExcluded) {
+      window.localStorage.removeItem("plausible_ignore");
+    } else {
+      window.localStorage.setItem("plausible_ignore", "true");
+    }
+    this.analyticsExcluded =
+      window.localStorage.getItem("plausible_ignore") === "true";
+  }
 }
