+++ packages/studio-web/src/app/app.module.ts
@@ -15,7 +15,8 @@ import { UploadComponent } from "./upload/upload.component";
 import { TextFormatDialogComponent } from "./text-format-dialog/text-format-dialog.component";
 import { NgxRAWebComponentModule } from "@readalongs/ngx-web-component";
 import { defineCustomElements } from "@readalongs/web-component/loader";
-import { StudioComponent, PrivacyDialog } from "./studio/studio.component";
+import { StudioComponent } from "./studio/studio.component";
+import { PrivacyDialog } from "./app.component";
 import { ErrorPageComponent } from "./error-page/error-page.component";
 import { EditorComponent } from "./editor/editor.component";
 import { SharedModule } from "./shared/shared.module";
