+++ packages/studio-web/src/app/studio/studio.component.html
@@ -11,9 +11,6 @@
     <section>
       <div class="container mb-2">
         <div>
-          <h1 i18n="Welcome message for app" id="welcome-header">
-            Welcome to ReadAlong Studio
-          </h1>
           <p i18n="Description of app">
             This is a tool to help you make your own interactive 'readalong'
             that highlights words as they are spoken. Have a look at
@@ -28,13 +25,8 @@
             to get a better idea of what these are.
           </p>
           <p i18n="Get started">
-            To get started making your own, read our
-            <a
-              style="color: #07c; cursor: pointer"
-              (click)="openPrivacyDialog()"
-              >privacy and data sovereignty policy</a
-            >, then take a tour by clicking on the tour button below, and follow
-            the steps below.
+            To get started making your own, click on the tour button below, and
+            follow the steps below.
           </p>
           <div class="row">
             <div class="col-3"></div>
